refactor: move api-routes.ts to src/config/apiRoutes.ts
Aligns with the other shared-constant modules (toolNames.ts, roles.ts,
pubsubChannels.ts) which already live under src/config/. Server-only
constants (workspace-paths.ts) stay in server/ because they depend on
Node APIs. All 23 import sites updated.
