+++ server/routes/plugins.ts
@@ -26,7 +26,7 @@ import {
   overwriteSpreadsheet,
   isSpreadsheetPath,
 } from "../utils/spreadsheet-store.js";
-import { API_ROUTES } from "../../src/api-routes.js";
+import { API_ROUTES } from "../../src/config/apiRoutes.js";
 
 const router = Router();
