+++ server/routes/todos.ts
@@ -28,7 +28,7 @@ import {
   type DispatchErrorResponse,
 } from "./dispatchResponse.js";
 
-import { API_ROUTES } from "../../src/api-routes.js";
+import { API_ROUTES } from "../../src/config/apiRoutes.js";
 
 const router = Router();
